Move Audit event storage to audit service
