+++ Audit-Service/pom.xml
@@ -38,8 +38,8 @@
         </dependency>
         <dependency>
             <groupId>io.github.arivanamin.scm</groupId>
-            <artifactId>storage-common</artifactId>
-            <version>${storage-common-module-version}</version>
+            <artifactId>outbox-storage</artifactId>
+            <version>${outbox-storage-module-version}</version>
         </dependency>
 
         <dependency>
