+++ Audit-Service/src/main/java/io/github/arivanamin/scm/backend/audit/application/beans/AuditBeansConfig.java
@@ -1,10 +1,11 @@
 package io.github.arivanamin.scm.backend.audit.application.beans;
 
+import io.github.arivanamin.scm.backend.audit.core.command.CreateAuditEventCommand;
+import io.github.arivanamin.scm.backend.audit.core.persistence.AuditEventStorage;
 import io.github.arivanamin.scm.backend.audit.core.query.ReadAuditEventByIdQuery;
 import io.github.arivanamin.scm.backend.audit.core.query.ReadAuditEventsQuery;
-import io.github.arivanamin.scm.backend.base.domain.audit.AuditEventStorage;
-import io.github.arivanamin.scm.backend.common.storage.audit.AuditEventRepository;
-import io.github.arivanamin.scm.backend.common.storage.audit.JpaAuditEventStorage;
+import io.github.arivanamin.scm.backend.audit.storage.AuditEventRepository;
+import io.github.arivanamin.scm.backend.audit.storage.JpaAuditEventStorage;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -24,5 +25,10 @@ class AuditBeansConfig {
     @Bean
     public ReadAuditEventByIdQuery readAuditEventByIdQuery (AuditEventStorage storage) {
         return new ReadAuditEventByIdQuery(storage);
+    }
+
+    @Bean
+    public CreateAuditEventCommand createAuditEventCommand (AuditEventStorage storage) {
+        return new CreateAuditEventCommand(storage);
     }
 }
