+++ Audit-Service/src/main/java/io/github/arivanamin/scm/backend/audit/core/exception/AuditEventNotFoundException.java
@@ -1,0 +1,8 @@
+package io.github.arivanamin.scm.backend.audit.core.exception;
+
+public class AuditEventNotFoundException extends RuntimeException {
+
+    public AuditEventNotFoundException () {
+        super("AuditEvent by the requested id not found");
+    }
+}
