+++ Audit-Service/src/main/java/io/github/arivanamin/scm/backend/audit/core/persistence/AuditEventStorage.java
@@ -1,5 +1,6 @@
-package io.github.arivanamin.scm.backend.base.domain.audit;
+package io.github.arivanamin.scm.backend.audit.core.persistence;
 
+import io.github.arivanamin.scm.backend.base.domain.audit.AuditEvent;
 import io.github.arivanamin.scm.backend.base.domain.dates.DateTimeRange;
 import io.github.arivanamin.scm.backend.base.domain.pagination.PaginatedResponse;
 import io.github.arivanamin.scm.backend.base.domain.pagination.PaginationCriteria;
