+++ Audit-Service/src/main/java/io/github/arivanamin/scm/backend/audit/storage/AuditEventRepository.java
@@ -1,4 +1,4 @@
-package io.github.arivanamin.scm.backend.common.storage.audit;
+package io.github.arivanamin.scm.backend.audit.storage;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
