+++ Audit-Service/src/main/resources/db/changelog/001-SCM-3-create-audit-events-table.xml
