+++ Audit-Service/src/main/resources/db/changelog/changelog-master.xml
@@ -5,5 +5,6 @@
     xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog
         http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-4.23.xsd">
 
-    <include file="SCM-3-create-audit-events-table.xml" relativeToChangelogFile="true" />
+    <include file="001-SCM-8-create-audit-events-outbox-table.xml" relativeToChangelogFile="true" />
+    <include file="001-SCM-3-create-audit-events-table.xml" relativeToChangelogFile="true" />
 </databaseChangeLog>
