+++ Audit-Service/src/test/java/io/github/arivanamin/scm/backend/audit/AuditEventTestingUtility.java
@@ -7,13 +7,15 @@ import org.instancio.Instancio;
 
 import java.time.LocalDateTime;
 
+import static io.github.arivanamin.scm.backend.base.domain.dates.TimestampHelper.toTimestampInMilliseconds;
+
 @NoArgsConstructor (access = AccessLevel.PRIVATE)
 public final class AuditEventTestingUtility {
 
     public static AuditEvent createSampleEvent (LocalDateTime recordedAt) {
         AuditEvent entity = Instancio.create(AuditEvent.class);
         entity.setId(null);
-        entity.setRecordedAt(recordedAt);
+        entity.setTimestamp(toTimestampInMilliseconds(recordedAt));
         return entity;
     }
 }
